requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.2.1
asyncpg==0.30.0
redis==5.2.1
python-dotenv==1.0.1
email-validator==2.2.0